+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Fix Dashboard wrong counters
 * Fixed icon path in development mode
+* Fixed window restoration on MacOS
 
 ### Changed